feat (airlines): Created util to keep specific logic outside view
